requirements.txt
# Week 1: Core Requirements
# Python version

# Core dependencies
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=0.24.0

# Advanced ML
pyyaml>=5.4.0
tqdm>=4.62.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Privacy & Evaluation
opacus>=0.15.0  # for differential privacy (Week 3)
tensorboard>=2.7.0

# Development tools
black>=21.11.0
flake8>=4.0.0
pytest>=6.2.0